Create project template (#12)
Add nix flake based project template. This allows downstream users to
have a development shell similar to pgzx and reuse some of our scripts.
The template create a small functioning hello world based extension
(without tests, thought).

The experience is still a little rough because Zig and Nix expect a
public repository. For now I made the template so that it works with a
local checkout.

---------

Co-authored-by: Tudor Golubenco <tudor.g@gmail.com>