Introduce Collections and tests
Introduce the `pgzx_unit` extension that is used to execute pgzx specific unit tests in Postgres.

The extensions entrypoint is `src/testing.zig`. Here we register all test suites that will later be run by the `./ci/run.sh` script. We basically use the same mechanism that we use for normal extensions code.

Other changes:
- Use code blocks in `build.zig` to better "isolate" doc build, test build, and the module definition.
- some fixes to ./ci/run.sh. Now we correctly fail early on errors and only print new logs (since beginning of our tests) on errors
- moved `pgzx.list` to `pgzx.collections.list`. We still have a `pgzx.list` alias for convenience.
- applied some fixes to the PointerList implementation to make sure that the constructors compile.
- initial implementation for slist support sneaked in somehow 🤷  

Improvements to the PointerList API, so to make it easier to actually create and use a postgres pointer list type from zig code while trying to reduce the risk of using the C API incorrectly.

- allows List in PointerList to be a null pointer. This is how empty lists are represented in postgres
- methods that not return a new list but modify the currently allocated list:
  - accept a pointer value to indicate that the current list is modified
  - do return void, because the current list is modified and we do not want to give the impression that a new list has been allocated
- updated iterator to allow initialization from an empty list (null pointer). No need to check if the list is null anymore.
- add some more delete methods

Introduce SList and DList types to support postgres intrusive lists.