Dynahash: init support for Postgres Hash tables. (#22)
Add typed wrappers for postgres HTAB type (`dynahash` hash tables that
can also be used with shared memory).

When using HTAB one confiures an 'Entry' that can be any kind of struct.
The prefix (first field) of the struct is supposed to hold the key. For
string keys this means that we have to configure a max string length.