Update env: docker and flakes (#47)
Update zig toolchain

Update Docker image: Ubuntu images now have a default user `ubuntu`. I
tried to update the Dockerfile to use the new user, but didn't get it
functioning correctly. Instead for now I opted to delete the `ubuntu`
user and keep the `dev` user an environment as is.