fdw: init submodule with helpers to implement foreign data wrappers (#52)
Requires #48, #49, #50

Implement `fdw.Options` related types and helper to validate options.

When implementing a foreign data wrapper one should implement a handler
that returns a list of callbacks and a validator function that validates
the settings FDW SQL statements will accept. So far the module only
implements helpers for the validator.