libpq: API changes
Introduce Options to configure if connections shall be created async or
not or if we should double check a connection after we have waited for
the connection to be established.
The new Options struct is used to combine the different connectX
functions into a single one.

Add more helpers to check for command completion: tryGetCommandOk,
waitCommandComplete, waitLastCommandComplete