elog: add XWithCause to include the cause in the error report (#62)
Sometimes we want to log an error after capturing another error.

Postgres errors and warnings are normally added to the error stack
(default size 5). The top entry of the error stack can be flushed via
`pg.EmitErrorReport`. Due to pgzx capturing errors and returning the
`PGStackError` we do not immediately call emit right now.

With this change we introduce `XWithCause` functions to elog, like
`LogWithCause`, `ErrorWithCause` and so on. These functions accept an
existing Zig error. In case the error given is `PGStackError` we will
call `pg.EmitErrorReport` to flush the original error. In case the error
is not `PGStackError` we print the the error message and append the
error code to the message like so: `<my error message>: <cause>`. This
pattern allows us to use the `WithCause` functions with errors we
captured using the Postgres API and Zig error codes using the same API.