ci: run scripts per example (#75)
Create a ci/run.sh script per example. Eventually we might want to
provide that functionality via the zig build system as extra tasks. But
there are a few things that I would like to change in the build helpers
that we have before adding test runners to it.

The main ci/run.sh script will build the examples in parallel now.