add missing file in zon file
The build system requires the `tools/gennodetags/main.zig` file, but
since it's not included in the `build.zig.zon` file, anyone who
adds `pgzx` as a dependency via `zig fetch` (or by specifying the
url) will not be able to build their extension. It will fail with a
`FileNotFound` error.